Add connect fn and integrate with tests
