+++ crates/proto/Cargo.toml
@@ -15,12 +15,14 @@ version.workspace      = true
 workspace = true
 
 [dependencies]
+anyhow           = { workspace = true }
 hex              = { version = "0.4" }
 miden-node-utils = { workspace = true }
 miden-objects    = { workspace = true }
 prost            = { workspace = true }
 thiserror        = { workspace = true }
 tonic            = { workspace = true }
+url              = { workspace = true }
 
 [dev-dependencies]
 proptest = { version = "1.5" }
