+++ crates/rpc/Cargo.toml
@@ -31,6 +31,7 @@ tonic-web        = { version = "0.12" }
 tower            = { workspace = true }
 tower-http       = { workspace = true, features = ["trace"] }
 tracing          = { workspace = true }
+url              = { workspace = true }
 
 [dev-dependencies]
 miden-node-rpc   = { workspace = true }
